functions/requirements.txt
azure-functions>=1.17.0
azure-identity>=1.15.0
openai>=1.12.0
psycopg2-binary>=2.9.9
pydantic>=2.6.0
pydantic-settings>=2.1.0
jsonschema>=4.21.0
